Refactor to package-by-feature structure
