+++ src/main/kotlin/io/github/nomisrev/profiles/ProfileRoutes.kt
@@ -1,12 +1,14 @@
 @file:Suppress("MatchingDeclarationName")
 
-package io.github.nomisrev.routes
+package io.github.nomisrev.profiles
 
 import arrow.core.raise.context.ensure
+import io.github.nomisrev.Api
 import io.github.nomisrev.MissingParameter
+import io.github.nomisrev.auth.JwtService
 import io.github.nomisrev.auth.jwtAuth
-import io.github.nomisrev.repo.UserPersistence
-import io.github.nomisrev.service.JwtService
+import io.github.nomisrev.route
+import io.github.nomisrev.users.UserPersistence
 import io.ktor.server.routing.Route
 import kotlinx.serialization.Serializable
 import opensavvy.spine.server.respond
