+++ src/main/kotlin/io/github/nomisrev/tags/TagPersistence.kt
@@ -1,5 +1,6 @@
-package io.github.nomisrev.repo
+package io.github.nomisrev.tags
 
+import io.github.nomisrev.articles.ArticleId
 import io.github.nomisrev.sqldelight.TagsQueries
 
 class TagPersistence(
