+++ src/test/kotlin/io/github/nomisrev/ValidationSpec.kt
@@ -2,18 +2,18 @@ package io.github.nomisrev
 
 import arrow.core.nonEmptyListOf
 import arrow.core.raise.either
-import io.github.nomisrev.repo.UserId
-import io.github.nomisrev.routes.ArticleResource
-import io.github.nomisrev.routes.ArticlesResource
-import io.github.nomisrev.routes.FeedLimit
-import io.github.nomisrev.routes.FeedOffset
-import io.github.nomisrev.routes.NewArticle
-import io.github.nomisrev.routes.NewComment
-import io.github.nomisrev.service.GetArticles
-import io.github.nomisrev.service.GetFeed
-import io.github.nomisrev.service.Login
-import io.github.nomisrev.service.RegisterUser
-import io.github.nomisrev.service.Update
+import io.github.nomisrev.articles.ArticleResource
+import io.github.nomisrev.articles.ArticlesResource
+import io.github.nomisrev.articles.FeedLimit
+import io.github.nomisrev.articles.FeedOffset
+import io.github.nomisrev.articles.GetArticles
+import io.github.nomisrev.articles.GetFeed
+import io.github.nomisrev.articles.NewArticle
+import io.github.nomisrev.articles.NewComment
+import io.github.nomisrev.users.Login
+import io.github.nomisrev.users.RegisterUser
+import io.github.nomisrev.users.Update
+import io.github.nomisrev.users.UserId
 import io.kotest.assertions.arrow.core.shouldBeLeft
 import io.kotest.assertions.arrow.core.shouldBeRight
 import io.kotest.core.spec.style.FreeSpec
