+++ src/test/kotlin/io/github/nomisrev/articles/ArticleRouteSpec.kt
@@ -1,13 +1,15 @@
-package io.github.nomisrev.routes
+package io.github.nomisrev.articles
 
 import arrow.core.raise.either
+import io.github.nomisrev.Api
+import io.github.nomisrev.Api.Article
+import io.github.nomisrev.Api.Article.feed
+import io.github.nomisrev.Api.Articles
+import io.github.nomisrev.Api.Articles.create
+import io.github.nomisrev.GenericErrorModel
 import io.github.nomisrev.articleFixture
-import io.github.nomisrev.routes.Api.Article
-import io.github.nomisrev.routes.Api.Article.feed
-import io.github.nomisrev.routes.Api.Articles
-import io.github.nomisrev.routes.Api.Articles.create
-import io.github.nomisrev.service.RegisterUser
 import io.github.nomisrev.userFixture
+import io.github.nomisrev.users.RegisterUser
 import io.github.nomisrev.withServer
 import io.kotest.assertions.arrow.core.shouldBeRight
 import io.kotest.core.spec.style.StringSpec
