+++ src/test/kotlin/io/github/nomisrev/tags/TagRouteSpec.kt
@@ -1,16 +1,17 @@
-package io.github.nomisrev.routes
+package io.github.nomisrev.tags
 
 import arrow.core.flatMap
 import arrow.core.raise.either
 import io.github.nefilim.kjwt.JWSHMAC512Algorithm
 import io.github.nefilim.kjwt.JWT
+import io.github.nomisrev.Api
+import io.github.nomisrev.Api.Tags
+import io.github.nomisrev.Api.Tags.list
 import io.github.nomisrev.articleFixture
-import io.github.nomisrev.repo.UserId
-import io.github.nomisrev.routes.Api.Tags
-import io.github.nomisrev.routes.Api.Tags.list
-import io.github.nomisrev.service.CreateArticle
-import io.github.nomisrev.service.RegisterUser
+import io.github.nomisrev.articles.CreateArticle
 import io.github.nomisrev.userFixture
+import io.github.nomisrev.users.RegisterUser
+import io.github.nomisrev.users.UserId
 import io.github.nomisrev.withServer
 import io.kotest.assertions.arrow.core.shouldBeRight
 import io.kotest.assertions.arrow.core.shouldBeSome
